+++ src/theme-src/webpack.config.js
@@ -1,5 +1,6 @@
 const path = require("path");
 const MiniCssExtractPlugin = require("mini-css-extract-plugin");
+const StyleLintPlugin = require("stylelint-webpack-plugin");
 const { CleanWebpackPlugin } = require("clean-webpack-plugin");
 
 module.exports = {
@@ -15,6 +16,10 @@ module.exports = {
     new CleanWebpackPlugin(),
     new MiniCssExtractPlugin({
       filename: "theme.css",
+    }),
+    new StyleLintPlugin({
+      files: "src/*.css",
+      fix: true,
     }),
   ],
   module: {